Now the time intervals show nicely.
maybe we could add a quick view of the event when it is clicked in the
calendar?
also, i need to make the events clickable in the first place and see how
can i hide the calendar while letting the month selector work. Or make
them appear on top of the calendar, whichever would be great.